# docs/workflow/workflow-page-api-spec.yaml
openapi: 3.0.0
info:
  title: 工作流详情页 API
  version: 1.0.0
  description: |
    工作流详情页（/workflow/:id）相关的所有后端 API 接口定义。

    ## 页面功能概述

    工作流详情页是工作流编辑器的核心页面，支持以下功能：
    - 工作流基本信息管理（名称、描述、触发配置）
    - 工作流画布编辑（节点、连线、关联）
    - 工作流执行与状态监控
    - 执行历史查看
    - 工作流发布/取消发布

    ## API 分类

    1. **WorkflowController** - 工作流基础管理
    2. **WorkflowExecutionController** - 工作流执行管理
    3. **WorkflowNodeController** - 节点管理
    4. **WorkflowConnectionController** - 连线管理
    5. **WorkflowValidationController** - 验证与辅助接口
servers:
  - url: http://localhost:8080
    description: 后端服务地址
paths:
...
  /api/workflow/{id}/execute:
    post:
      summary: 执行工作流
      description: |
        触发工作流的执行。

        ## 执行流程
        1. 后端创建一条执行记录（Execution）
        2. 异步执行工作流中的各个节点
        3. 返回执行记录ID，前端可通过此ID轮询执行状态

        ## 使用场景
        - 用户在工作流详情页点击"运行"按钮
        - 通过输入面板配置输入参数后执行
      operationId: executeWorkflow
      tags:
        - WorkflowExecutionController
      parameters:
        - name: id
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: triggeredBy
          in: query
          description: 触发人/触发来源
          schema:
            type: string
            example: "manual"
      requestBody:
        description: 工作流输入参数（JSON格式）
        content:
          application/json:
            schema:
              type: object
              description: 键值对形式的工作流输入参数
              example:
                input_text: "测试输入文本"
                dataset_id: 123
      responses:
        '200':
          description: 执行记录ID
          content:
            application/json:
              schema:
                type: integer
                format: int64
                description: 执行记录ID，用于后续查询执行状态
                example: 456
  /api/workflow/execution/{executionId}:
    get:
      summary: 获取执行记录详情
      description: |
        根据执行记录ID获取详细的执行信息，包括：
        - 执行状态（RUNNING/COMPLETED/FAILED/ABORTED）
        - 输入数据
        - 输出数据
        - 开始时间和结束时间
        - 错误信息（如果失败）

        ## 使用场景
        - 执行工作流后，前端轮询此接口获取执行状态
        - 在执行日志面板中显示实时执行进度
      operationId: getExecutionDetail
      tags:
        - WorkflowExecutionController
      parameters:
        - name: executionId
          in: path
          required: true
          description: 执行记录ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 执行记录详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
  /api/workflow/execution/uuid/{uuid}:
    get:
      summary: 根据UUID获取执行记录
      description: |
        通过执行记录的UUID获取执行详情。

        ## 使用场景
        - 外部系统通过UUID查询执行结果
        - Webhook回调时验证执行状态
      operationId: getExecutionByUuid
      tags:
        - WorkflowExecutionController
      parameters:
        - name: uuid
          in: path
          required: true
          description: 执行记录UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 执行记录详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
  /api/workflow/execution/{executionId}/abort:
    post:
      summary: 中止执行
      description: |
        中止正在执行中的工作流。

        ## 使用场景
        - 用户点击"停止执行"按钮
        - 执行超时时自动中止
      operationId: abortExecution
      tags:
        - WorkflowExecutionController
      parameters:
        - name: executionId
          in: path
          required: true
          description: 执行记录ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 中止成功
  /api/workflow/{workflowId}/executions:
    get:
      summary: 获取工作流执行历史
      description: |
        分页获取指定工作流的所有执行记录历史。

        ## 使用场景
        - 在执行历史面板中查看历史执行记录
        - 分析工作流的执行情况和成功率
      operationId: getWorkflowExecutions
      tags:
        - WorkflowExecutionController
      parameters:
        - name: workflowId
          in: path
          required: true
          description: 工作流ID
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: 页码（从0开始）
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 每页大小
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 执行记录分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionPageResponse'
...
  /api/workflow/execution/{executionId}/outputs:
    get:
      summary: 获取执行输出参数
      description: |
        获取工作流执行完成后的输出参数。

        ## 使用场景
        - 执行完成后，在结果面板中显示输出
        - 获取工作流的执行结果数据
      operationId: getExecutionOutputs
      tags:
        - WorkflowExecutionController
      parameters:
        - name: executionId
          in: path
          required: true
          description: 执行记录ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 输出参数
          content:
            application/json:
              schema:
                type: object
                description: 键值对形式的输出参数

  # ==================== 节点管理 (WorkflowNodeController) ====================
...
components:
  schemas:
...
    ExecutionResponse:
      type: object
      description: 执行记录响应
      properties:
        id:
          type: integer
          format: int64
          description: 执行记录ID
          example: 456
        uuid:
          type: string
          description: 执行UUID
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        workflowId:
          type: integer
          format: int64
          description: 工作流ID
          example: 30
        status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED, ABORTED]
          description: 执行状态
          example: "COMPLETED"
        inputData:
          type: string
          description: 输入数据（JSON格式）
          example: '{"input_text": "测试输入"}'
        outputData:
          type: string
          description: 输出数据（JSON格式）
          example: '{"result": "处理完成"}'
        errorMessage:
          type: string
          description: 错误信息（失败时）
          example: null
        startedAt:
          type: string
          format: date-time
          description: 开始时间
          example: "2024-04-07T10:30:00Z"
        finishedAt:
          type: string
          format: date-time
          description: 结束时间
          example: "2024-04-07T10:31:30Z"
    ExecutionPageResponse:
      type: object
      description: 执行记录分页响应
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionResponse'
        totalElements:
          type: integer
          description: 总记录数
          example: 100
        totalPages:
          type: integer
          description: 总页数
          example: 10
        number:
          type: integer
          description: 当前页码（从0开始）
          example: 0
        size:
          type: integer
          description: 每页大小
          example: 10

    # ==================== 验证相关 ====================
...